~updated database ~updated exception toast ~fixed a protocol bug

+++ app/src/main/java/com/third/lhat/database/model/Server.kt
@@ -1,18 +1,16 @@
 package com.third.lhat.database.model
 
 import androidx.room.Entity
-import androidx.room.Index
 import androidx.room.PrimaryKey
 import androidx.room.TypeConverters
-import com.third.lhat.database.converter.ListConverter
 import com.third.lhat.database.converter.LocalDateTimeConverter
 import java.time.LocalDateTime
 
 @Entity
 @TypeConverters(LocalDateTimeConverter::class)
 data class Server(
-    @PrimaryKey(autoGenerate = true) val serverId: Int,
-    val serverUserId: Int,
+    @PrimaryKey(autoGenerate = true) val serverId: Long,
+    val serverUserId: Long,
     val address: String,
     val port: Int,
     val lastLogin: LocalDateTime
